pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bcrypt==4.3.0",
    "fastapi[standard]>=0.128.1",
    "grpcio>=1.76.0",
    "passlib[bcrypt]>=1.7.4",
    "protobuf>=6.33.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.46",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
]
